Move common components to separate dir
